Refactor createPaymentRequestReweighUpdate
