+++ src/constants/MoveHistory/EventTemplates/UpdateReweigh/createPaymentRequestReweighUpdate.jsx
@@ -1,5 +1,6 @@
+import React from 'react';
+
 import o from 'constants/MoveHistory/UIDisplay/Operations';
-import d from 'constants/MoveHistory/UIDisplay/DetailsTypes';
 import a from 'constants/MoveHistory/Database/Actions';
 import t from 'constants/MoveHistory/Database/Tables';
 
@@ -7,9 +8,10 @@ export default {
   action: a.INSERT,
   eventName: o.updateReweigh,
   tableName: t.payment_requests,
-  detailsType: d.STATUS,
   getEventNameDisplay: ({ changedValues }) => `Created payment request ${changedValues?.payment_request_number}`,
-  getStatusDetails: () => {
-    return 'Pending';
-  },
+  getDetails: () => (
+    <>
+      <b>Status</b>: Pending
+    </>
+  ),
 };
