+++ src/constants/MoveHistory/EventTemplates/UpdateReweigh/createPaymentRequestReweighUpdate.test.jsx
@@ -1,15 +1,18 @@
+import { render, screen } from '@testing-library/react';
+
 import getTemplate from 'constants/MoveHistory/TemplateManager';
-import e from 'constants/MoveHistory/EventTemplates/createPaymentRequestReweighUpdate';
 
 describe('when given a payment request is created through reweigh', () => {
-  const item = {
+  const historyRecord = {
     action: 'INSERT',
     eventName: 'updateReweigh',
     tableName: 'payment_requests',
   };
   it('correctly matches the Request shipment reweigh event', () => {
-    const result = getTemplate(item);
-    expect(result).toMatchObject(e);
-    expect(result.getStatusDetails(item)).toEqual('Pending');
+    const template = getTemplate(historyRecord);
+
+    render(template.getDetails(historyRecord));
+    expect(screen.getByText('Status')).toBeInTheDocument();
+    expect(screen.getByText(': Pending')).toBeInTheDocument();
   });
 });
